+++ cmd/message/list.go
@@ -6,6 +6,7 @@ import (
 	"text/tabwriter"
 
 	"github.com/prolific-oss/cli/client"
+	"github.com/prolific-oss/cli/config"
 	"github.com/prolific-oss/cli/ui"
 	"github.com/spf13/cobra"
 )
@@ -96,5 +97,9 @@ func renderMessages(c client.API, opts ListOptions, w io.Writer) error {
 		}
 	}
 
-	return tw.Flush()
+	_ = tw.Flush()
+
+	fmt.Fprintf(w, "\nView messages in the application: %s/messages/inbox\n", config.GetApplicationURL())
+
+	return nil
 }